feat: migration added to update osm_id type from int to bigint (#2552)
* feat: migration added to update osm_id type from int to bigint

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* refactor: move migrations dir from backend to src (for frontend + backend)

* build: move migrations to root src dir for both frontend / backend migrations

---------

+++ src/migrations/init/shared/2-tables.sql
@@ -27,7 +27,7 @@ CREATE TABLE public.odk_entities (
     status public.entitystate NOT NULL,
     project_id integer NOT NULL,
     task_id integer,
-    osm_id integer,
+    osm_id bigint,
     submission_ids character varying,
     is_new boolean DEFAULT false,
     -- This a javarosa geom string
